requirements.txt
fastapi
uvicorn
pillow
imageio
numpy
python-multipart
aiohttp
